Begun transition to LibGDX; implemented screens and scenegraph.
There's still a lot more to implement before the game is as it was under
Slick2d. So far we have a main menu, a game screen which uses a scene graph
for entity management, and the beginnings of a UI framework. Still to
implement: tile rendering; NPC movement, collision checking; reading data
from SQLite. After that, we start on NPC actions, menus, saving, combat,
items, etc. Phew...